:sparkles: Use rich text rendering instead of Chrome for HTML clipboard rendering
