+++ app/src/commonMain/kotlin/com/crosspaste/db/task/TaskType.kt
@@ -7,6 +7,7 @@ object TaskType {
     const val DELETE_PASTE_TASK = 1
     const val PULL_FILE_TASK = 2
     const val CLEAN_PASTE_TASK = 3
+    @Deprecated("not used anymore")
     const val HTML_TO_IMAGE_TASK = 4
     const val PULL_ICON_TASK = 5
     const val RTF_TO_IMAGE_TASK = 6
