+++ app/src/desktopMain/kotlin/com/crosspaste/ui/paste/detail/RtfDetailView.kt
@@ -1,70 +1,82 @@
 package com.crosspaste.ui.paste.detail
 
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.text.style.TextOverflow
 import com.crosspaste.i18n.GlobalCopywriter
 import com.crosspaste.info.PasteInfos.DATE
 import com.crosspaste.info.PasteInfos.REMOTE
 import com.crosspaste.info.PasteInfos.SIZE
 import com.crosspaste.info.PasteInfos.TYPE
-import com.crosspaste.paste.item.PasteItem
-import com.crosspaste.paste.item.TextPasteItem
+import com.crosspaste.paste.item.PasteText
+import com.crosspaste.paste.item.RtfPasteItem
+import com.crosspaste.path.UserDataPathProvider
+import com.crosspaste.ui.base.UISupport
+import com.crosspaste.ui.paste.GenerateImageView
 import com.crosspaste.ui.paste.PasteDataScope
-import com.crosspaste.ui.theme.AppUISize.small3X
-import com.crosspaste.ui.theme.DesktopAppUIFont.detailPasteTextStyle
 import com.crosspaste.utils.DateUtils
 import com.crosspaste.utils.getFileUtils
 import org.koin.compose.koinInject
 
 @Composable
-fun PasteDataScope.PasteTextDetailView(onDoubleClick: () -> Unit) {
+fun PasteDataScope.RtfDetailView(onDoubleClick: () -> Unit) {
     val copywriter = koinInject<GlobalCopywriter>()
-    val fileUtils = getFileUtils()
+    val uiSupport = koinInject<UISupport>()
+    val userDataPathProvider = koinInject<UserDataPathProvider>()
 
-    val textPasteItem = getPasteItem(TextPasteItem::class)
-    val text = textPasteItem.text
-    val pasteItem = textPasteItem as PasteItem
+    val fileUtils = getFileUtils()
+    val rtfPasteItem = getPasteItem(RtfPasteItem::class)
+
+    val filePath by remember(pasteData.id) {
+        mutableStateOf(
+            rtfPasteItem.getRenderingFilePath(
+                pasteData.getPasteCoordinate(),
+                userDataPathProvider,
+            ),
+        )
+    }
 
     PasteDetailView(
         detailView = {
-            Row(
+            val horizontalScrollState = rememberScrollState()
+            val verticalScrollState = rememberScrollState()
+            GenerateImageView(
                 modifier =
                     Modifier
                         .fillMaxSize()
+                        .horizontalScroll(horizontalScrollState)
+                        .verticalScroll(verticalScrollState)
                         .pointerInput(Unit) {
                             detectTapGestures(
+                                onTap = {
+                                    uiSupport.openRtf(pasteData)
+                                },
                                 onDoubleTap = {
                                     onDoubleClick()
                                 },
                             )
-                        }.padding(small3X),
-            ) {
-                Text(
-                    text = text,
-                    modifier =
-                        Modifier
-                            .fillMaxSize()
-                            .verticalScroll(rememberScrollState()),
-                    overflow = TextOverflow.Ellipsis,
-                    style = detailPasteTextStyle,
-                )
-            }
+                        },
+                imagePath = filePath,
+                text = pasteData.getPasteItem(PasteText::class)?.text ?: rtfPasteItem.getText(),
+                preview = false,
+                alignment = Alignment.TopStart,
+            )
         },
         detailInfoView = {
             PasteDetailInfoView(
                 items =
                     listOf(
-                        PasteDetailInfoItem(TYPE, copywriter.getText("text")),
-                        PasteDetailInfoItem(SIZE, fileUtils.formatBytes(pasteItem.size)),
+                        PasteDetailInfoItem(TYPE, copywriter.getText("rtf")),
+                        PasteDetailInfoItem(SIZE, fileUtils.formatBytes(rtfPasteItem.size)),
                         PasteDetailInfoItem(REMOTE, copywriter.getText(if (pasteData.remote) "yes" else "no")),
                         PasteDetailInfoItem(
                             DATE,
