+++ app/src/desktopMain/kotlin/com/crosspaste/ui/paste/detail/UrlDetailView.kt
@@ -1,83 +1,67 @@
 package com.crosspaste.ui.paste.detail
 
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.text.style.TextOverflow
 import com.crosspaste.i18n.GlobalCopywriter
 import com.crosspaste.info.PasteInfos.DATE
 import com.crosspaste.info.PasteInfos.REMOTE
 import com.crosspaste.info.PasteInfos.SIZE
 import com.crosspaste.info.PasteInfos.TYPE
-import com.crosspaste.paste.item.HtmlPasteItem
-import com.crosspaste.paste.item.PasteText
-import com.crosspaste.path.UserDataPathProvider
+import com.crosspaste.paste.item.UrlPasteItem
 import com.crosspaste.ui.base.UISupport
-import com.crosspaste.ui.paste.GenerateImageView
 import com.crosspaste.ui.paste.PasteDataScope
+import com.crosspaste.ui.theme.AppUIFont.pasteUrlStyle
+import com.crosspaste.ui.theme.AppUISize.small3X
 import com.crosspaste.utils.DateUtils
 import com.crosspaste.utils.getFileUtils
 import org.koin.compose.koinInject
 
 @Composable
-fun PasteDataScope.HtmlToImageDetailView(onDoubleClick: () -> Unit) {
+fun PasteDataScope.UrlDetailView(onDoubleClick: () -> Unit) {
     val copywriter = koinInject<GlobalCopywriter>()
     val uiSupport = koinInject<UISupport>()
-    val userDataPathProvider = koinInject<UserDataPathProvider>()
-
     val fileUtils = getFileUtils()
-
-    val htmlPasteItem = getPasteItem(HtmlPasteItem::class)
-
-    val filePath by remember(pasteData.id) {
-        mutableStateOf(
-            htmlPasteItem.getRenderingFilePath(
-                pasteData.getPasteCoordinate(),
-                userDataPathProvider,
-            ),
-        )
-    }
+    val urlPasteItem = getPasteItem(UrlPasteItem::class)
+    val url = urlPasteItem.url
 
     PasteDetailView(
         detailView = {
-            val horizontalScrollState = rememberScrollState()
-            val verticalScrollState = rememberScrollState()
-            GenerateImageView(
+            Row(
                 modifier =
                     Modifier
                         .fillMaxSize()
-                        .horizontalScroll(horizontalScrollState)
-                        .verticalScroll(verticalScrollState)
                         .pointerInput(Unit) {
                             detectTapGestures(
                                 onTap = {
-                                    uiSupport.openHtml(pasteData.id, htmlPasteItem.html)
+                                    uiSupport.openUrlInBrowser(urlPasteItem.url)
                                 },
                                 onDoubleTap = {
                                     onDoubleClick()
                                 },
                             )
-                        },
-                imagePath = filePath,
-                text = pasteData.getPasteItem(PasteText::class)?.text ?: htmlPasteItem.getText(),
-                preview = false,
-                alignment = Alignment.TopStart,
-            )
+                        }.padding(small3X),
+            ) {
+                Text(
+                    text = url,
+                    modifier = Modifier.fillMaxSize(),
+                    overflow = TextOverflow.Ellipsis,
+                    style = pasteUrlStyle,
+                )
+            }
         },
         detailInfoView = {
             PasteDetailInfoView(
                 items =
                     listOf(
-                        PasteDetailInfoItem(TYPE, copywriter.getText("html")),
-                        PasteDetailInfoItem(SIZE, fileUtils.formatBytes(htmlPasteItem.size)),
+                        PasteDetailInfoItem(TYPE, copywriter.getText("link")),
+                        PasteDetailInfoItem(SIZE, fileUtils.formatBytes(urlPasteItem.size)),
                         PasteDetailInfoItem(REMOTE, copywriter.getText(if (pasteData.remote) "yes" else "no")),
                         PasteDetailInfoItem(
                             DATE,
